elastic_utils refactor
